#1345: split server/handlers.rs handle_stream into per-verb modules
`userspace-dp/src/server/handlers.rs` was 461 LOC dominated by a
single 415-LOC `handle_stream()` function — one giant
`match request.request_type.as_str()` dispatcher over 19 wire
verbs. That body is >4x the docs/engineering-style.md soft cap
(100 LOC) and >2x the Tier-1 hard cap (200 LOC), and nobody could
review it end-to-end.

This is pure code motion. The file is replaced with a directory
module `server/handlers/`:

- `handlers/mod.rs` (181 LOC) keeps the I/O setup, decode,
  `state.lock()` critical section, dispatch shell, post-match
  status attach, post-lock `write_state`, and response write.
- 12 per-verb files contain the substantive match-arm bodies:
  snapshot.rs (apply_snapshot, bump_fib_generation),
  forwarding.rs, ha.rs, neighbors.rs, queue.rs, binding.rs,
  inject_packet.rs, sync_session.rs, session_deltas.rs,
  export.rs (owner_rg + all), rebind.rs, stop_workers.rs.
- Trivial arms (≤5 LOC bodies) stay inline in mod.rs's match:
  ping|status (no-op), update_fabrics (single helper call),
  clear_policy_counters (3-line body), shutdown (3-line body —
  also the ONLY arm that touches the `running: Arc<AtomicBool>`,
  so keeping it inline avoids threading the Arc through any
  per-verb file), catch-all (error path).

Per-verb handlers take `&mut ServerState`, the consumed
`Option<T>` request field by value (no clones — preserves
today's `if let Some(...) = request.foo` move semantics
byte-identical), and `&mut ControlResponse` /
`&mut bool persist_state` parameters only where they're written.

Wire protocol unchanged — verb strings, request fields,
response fields are all identical. `eprintln!` log strings are
preserved byte-for-byte. `refresh_status` call sites stay where
they were (16 eager in-arm + 1 post-match gated by
`!suppress_status`). `persist_state` stays sticky.

Plan + adversarial-review history:
- docs/pr/1345-server-handlers-split/plan.md (v1 → v2 → v3)
- AGY r1 (v1): PLAN-KILL on over-fragmentation + clone cost
- AGY r2 (v2): PLAN-NEEDS-MINOR (helper-call substitution)
- Codex r2 (on v1): PLAN-NEEDS-MAJOR; findings all addressed in
  v2/v3 (HandlerCtx dropped, clones eliminated, refresh_status
  count corrected from 17 to 16, lock invariant restated).
- Codex r3 (on v3): PLAN-READY.

Test status:
- cargo build clean (14 warnings, all pre-existing).
- cargo test --release: 1487 tests pass; the lone failure
  `snat_contract_documents_current_fail_closed_runtime` is a
  pre-existing master-baseline failure confirmed by running
  the same test against `origin/master` (936b076d) — the test
  checks docs/userspace-dataplane-gaps.md for "fail-closed",
  which is absent on master. NOT introduced by this PR.
- 5x flake check on apply_snapshot e2e tests (the only
  handle_stream e2e coverage in main_tests.rs): 5/5 pass.
- go test ./...: all 30+ packages clean.

Closes #1345
