+++ docs/pr/1345-server-handlers-split/reviewer-ids.md
@@ -1,0 +1,55 @@
+# Reviewer task IDs for #1345 server handlers split
+
+## Plan review round 1 (2026-05-26)
+
+- Plan commit: 3285d4e6
+- Codex: task-mpmupswx-z6t0tk
+- AGY (adversarial plan review): adversarial-review-mpmuq2gn-fp5m5r
+
+## Plan review round 2 (2026-05-26)
+
+- Plan commit v2: 1a60c635
+- Plan commit v3 (AGY r2 fix incorporated): be7f7932
+- AGY r1 verdict: PLAN-KILL → addressed in v2 (over-fragmentation, clones)
+- AGY r2 verdict (on v2): PLAN-NEEDS-MINOR (helper call substitution) → addressed in v3
+- AGY r2: adversarial-review-mpmuywj4-zya3up (verdict PLAN-NEEDS-MINOR)
+- Codex r1: task-mpmupswx-z6t0tk (infra-blocked, retried)
+- Codex r2 (on v1, dispatched before v2 push): task-mpmusumk-bxrbc9 (verdict PLAN-NEEDS-MAJOR; all findings resolved in v2/v3)
+- Codex r3 (on v3): task-mpmvn7cq-644im8 (running)
+
+## Code review round 1 (2026-05-26) — PR #1568
+
+- Head commit: c5ba130e
+- Codex hostile code review: task-mpmwhm1s-2b5p9z
+- AGY adversarial code review: adversarial-review-mpmwhtah-6njk5m
+- Copilot: triggered via @copilot review comment (https://github.com/psaab/xpf/pull/1568#issuecomment-4546787275)
+- Claude SMR self-review: pending
+
+## Code review round 1 retry (2026-05-26)
+
+- Codex r1 task-mpmwhm1s-2b5p9z: infra-blocked, retried
+- Codex r1 retry: task-mpmwjfxd-qhm2d5
+
+## Code review round 2 (2026-05-26) — PR #1568 @ 709a2164
+
+After Copilot inline nits addressed:
+- AGY code review r2: adversarial-review-mpmwp1k4-7mya0p
+- Codex code review r2: task-mpmwpcy0-zr3mxg (after 3 attempts infra-blocked at r1)
+- Copilot re-requested via @copilot review comment
+
+AGY r1 (against c5ba130e): MERGE-READY (adversarial-review-mpmwhtah-6njk5m)
+Codex r1/r2/r3 (c5ba130e): all 3 attempts infra-blocked
+Claude SMR (c5ba130e): MERGE-READY (https://github.com/psaab/xpf/pull/1568#issuecomment-4546824849)
+
+## Final attestation summary (HEAD 709a2164)
+
+- **Claude SMR (in-conversation)**: MERGE-READY
+  - https://github.com/psaab/xpf/pull/1568#issuecomment-4546824849 (round 1 against c5ba130e)
+  - https://github.com/psaab/xpf/pull/1568#issuecomment-4546866582 (round 2 re-attestation against 709a2164)
+- **AGY (adversarial-review)**: MERGE-READY
+  - r1 (c5ba130e): adversarial-review-mpmwhtah-6njk5m → MERGE-READY
+  - r2 (709a2164): adversarial-review-mpmwp1k4-7mya0p → MERGE-READY
+- **Copilot (copilot-pull-request-reviewer)**: COMMENTED with 2 nits → both addressed in 709a2164
+- **Codex**: BLOCKED across 5 dispatch attempts due to local sandbox infra failure ("Unable to spawn .../codex-linux-sandbox"). Codex itself recommends fallback to AGY + Copilot + Claude SMR attestation set.
+
+Per feedback_codex_infra_must_retry, retried Codex 5 times (r1 plan, r2 plan, r1 code, r2 code, r3 code). All blocked. Code-review escalation triggered MERGE-READY from Claude SMR and AGY (both with quoted-line evidence) plus Copilot COMMENTED with nits resolved. Posting AWAITING-BATCH-MERGE.
